FinalProjectPOC/demo/requirements.txt
# Demo dependencies for PostgreSQL to StarRocks Sync

# Core dependencies
psycopg2-binary==2.9.7
sqlalchemy==2.0.19
pandas==2.0.3
numpy==1.24.3

# Configuration and utilities
pydantic==2.1.1
pyyaml==6.0.1
python-dotenv==1.0.0
click==8.1.7
rich==13.4.2

# Logging
structlog==23.1.0

# HTTP client for health checks
requests==2.31.0

# Database drivers for real sync
psycopg2-binary==2.9.7
PyMySQL==1.1.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1
